Refactored code into separate modules Made program master-slave to run on 2 RPIs.
