+++ SlaveDataManager.py
@@ -1,7 +1,7 @@
-class DataManager(object):
+class SlaveDataManager(object):
     def __init__(self, handle):
         self.handle = handle
-        # TODO pull data from somewhere
+        # TODO accept commands from master
         self.power = 2.5
         self.speed = 15
         self.capacity = 88
@@ -17,8 +17,8 @@ class DataManager(object):
         handle.power_display.update_value(self.power)
         handle.speed_display.update_value(self.speed)
         handle.capacity_display.update_value(self.capacity)
-        handle.solar_display.update_value(self.solar_power)
-        handle.efficiency_display.update_value(self.efficiency)
+        handle.solar_poewr_displaty.update_value(self.solar_power)
+        handle.solar_poewr_displaty.update_value(self.efficiency)
         handle.range_display.update_value(self.range)
         handle.batt_temp_display.update_value(self.battery_temp)
         handle.motor_temp_display.update_value(self.motor_temp)
